Overhauled packet obfuscation handling
Switched it from packet_db to source side defines. This is the first step of making multiple packet versions on the same server unuseable. We decided to go into that direction, because it only gives us more problems and headaches if we feature this any longer.

Thanks to the hercules crew for this design idea.

At last a very big thank you to Yommy, Shakto, Themon and everyone else who might have contributed keys.